fix: reviewer feedback
